requirements.txt
# requirements.txt
pandas>=1.5.0
numpy>=1.21.0
PyYAML>=6.0
openpyxl>=3.0.0